refactor: deduplicate parse functions into core/parse-utils.js
Moved 9 duplicated functions from parse.js and worker-parse.js into
core/parse-utils.js (runtime-neutral utilities that work in both
main thread and web worker contexts).

Eliminated 447 lines of duplicate code.
